feat: add patient suggestion when creating appointment
